+++ agents-docs/navigation.ts
@@ -17,7 +17,6 @@ export default {
       pages: [
         'typescript-sdk/agent-settings',
         'typescript-sdk/models',
-        'typescript-sdk/memory',
         'typescript-sdk/agent-relationships',
         {
           group: 'Tools',
@@ -35,6 +34,7 @@ export default {
             'typescript-sdk/credentials/env-aware-credentials',
           ],
         },
+        'typescript-sdk/memory',
         'typescript-sdk/headers',
         'typescript-sdk/context-fetchers',
         'authentication',
@@ -159,6 +159,7 @@ export default {
             'tutorials/mcp-servers/overview',
             'tutorials/mcp-servers/native-mcp-servers',
             'tutorials/mcp-servers/composio-mcp-servers',
+            'tutorials/mcp-servers/gram',
             'tutorials/mcp-servers/custom-mcp-servers',
           ],
         },