+++ agents-docs/src/components/sidebar/index.tsx
@@ -1,3 +1,4 @@
+import { YoutubeIcon } from '@/components/brand-icons';
 import { List } from '@/components/sidebar/list';
 import { Sidebar, SidebarFooter } from '@/components/ui/sidebar';
 import { SLACK_URL } from '@/lib/constants';
@@ -25,6 +26,11 @@ const footerLinks = [
     href: 'https://twitter.com/inkeep',
     iconName: 'FaXTwitter',
     label: 'X (Twitter)',
+  },
+  {
+    href: 'https://www.youtube.com/@inkeep-ai',
+    icon: <YoutubeIcon />,
+    label: 'Inkeep on YouTube',
   },
 ];